Updated openstack/openstack Project: openstack/python-keystoneclient  f9be84a566fc91e20a2fc7eaa7d4faf2c7b064ec
Allow fetching user_id/project_id from auth

This would ideally not be required however when building certain URLs
the current user_id is needed. And when communicating with certain
services we need to have access to the current project id. It seems
better to allow plugins to give up the information if they have it than
do various hacks to try and get it from them.

Change-Id: Ib61b0628702806268be623a9987a922a60b04165
Closes-Bug: #1364724