Updated openstack/openstack Project: openstack/neutron  acfbb8636a357ddece1cd37f445625a0f97aaf8c
Provide doc string pointers for the dhcp agent rpc api

There is an rpc interface defined for the Neutron plugin to be able to
execute methods in the DHCP agent.  Provide docstring pointers in the
client and server side that tells you where to find the other side of
the interface.

No namespace usage is needed here.  This API is the only one exposed
via the DHCP agent, so the default namespace used now is fine.

The DhcpAgent class was updated to explicitly define the
messaging.Target().  Previously it was using the equivalent one
defined in the Manager base class.  Having it specified here makes it
more obvious that this is an rpc endpoint, and also provides the
obvious place that must have the version updated if the interface is
changed.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I4a6eb8dacb9ba01f329a5d5961dc0e0ee6f780ba