Updated openstack/openstack Project: openstack/neutron  436df515900b739476ba167e7043c8fdc1c901d3
Scope metadata rpc api using a messaging namespace

This patch does a couple of things.  First it adds docstrings to the
client/server pair of the rpc interface used by the metadata service
to make calls back into the Neutron server.  The docs tell you where
the other side of the interface is found in the code, and where docs
are that give more info on the rules for changing them.

The second thing done in this patch is to scope this interface using a
messaging namespace.  Right now some plugins expose several interfaces
via the default namespace.  This effectively means they are a single
API and should be managed with a single version stream.  It's much
more managable to just treat these as separate interfaces and this
change makes that explicit and functionally true.  Now when a method
is invoked, the only classes considered for handling that request will
be ones marked with the right namespace.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I3330229bf85b01d50c90e9ca064ae5e0fae83509