Updated openstack/openstack Project: openstack/neutron  cd0cf84ee39fef9d2dffd08523cd0e66a2c879ec
Move metadata agent entry to its own file

Break main() and configs out of agent logic. This makes metadata
consistent with the L3 and DHCP agents [1, 2].

Along with [3], this is done to better delineate what the agent module
inner boundaries are about. More to follow to finish off the cleanup.

[1] https://review.openstack.org/#/c/145979/
[2] https://review.openstack.org/#/c/146024/
[3] https://review.openstack.org/#/c/146151/

Change-Id: I2842f7c21db6f6dabdd2549f91a9001220454c22