Updated openstack/openstack Project: openstack/neutron  f83b08abf59a915206e93051af96bdcc5b7724c2
Confusing message deleting default security group

Trying to delete default security group from a non-admin
user gives an error "Removing default security group not
allowed. (HTTP 400).
But this message is not clear as the user does not get to
know why is he not allowed to remove the default security
group.

So updated the error message to give a more optimal message
to the user.

Change-Id: I25744b854371208d68a5cbdcf83079db7c54f96f
Closes-Bug: #1403743