Updated openstack/openstack Project: openstack/cinder  75214158021d55a182b79383e9e202478a018522
Clean up QoSSpecManageApiTest setup

There is currently unnecessary work being performed in setUp
for QoSSpecManageApiTest and confusing comments. This cleans
that up to avoid confusion.

Test was explicitly calling a reset on a fake, but the cleanup
for each test will do this as well. Relatively harmless, but
this makes the code cleaner.

Change-Id: I2671fc0a186acb75767d9adb1edeb9ad705fc958