Updated openstack/openstack Project: openstack/swift  d11de3753cf9fc4f06d822158b2cb5b9e2f34c3d
Fix slo constraint according to the error message

This patch allows to create a slo manifest when the manifest
includes only one segment less than min_segment_size.

When putting a manifest for slo with only one segment less than
min_segment_size, it will fail as EntityTooSmall with the message
"Each segment, except the last, must be at least min_segment_size
bytes." This behavior is different from the message because when
there is only one segment, the segment is absolutely the last
segment.

Change-Id: I8f0203afe55536207a41e1267128f8378f1ba15f