Updated openstack/openstack Project: openstack/ceilometer  de29b781196b3bde3a34699521a8929355d177a1
Catch exception when evaluate single alarm

Currently, if we raise exception when evaluate alarms, the outside
method will not catch it, so the higher _evaluate_assigned_alarms
catches it, but this will stop evaluating the rest alarms. Alarm
should be evaluated no matter whether other alarms succeed or not.

This patch adds a try...except block in _evaluate_alarm, it will
log exception when it is raised, and the higher method can move
on next alarm.

Change-Id: Id95865c51dc4ffe2d7089a3ff4fa5b73bd93ae76
Closes-Bug: #1408620