Updated openstack/openstack Project: openstack/swift  a2922a896c04687178ed107756a22edc95fe909c
Test that SLO disallows too small first segment if other segments

SLO allows the first segment to be less than min_segment_size if
it is the only segment. Current tests verify that a single small
segment is allowed, and that multiple small segments are disallowed.
This patch adds a test to verify that SLO will disallow a manifest
with a small first segment followed by another correctly sized
segment.

Change-Id: I920c0aee38e4e16c49bd84a3b772308a00794fa7