Updated openstack/openstack Project: openstack/oslo.db  b84d921ac8c29dadd4f5627dc7a78b2c01b7cdc3
Make sure sort_key_attr is QueryableAttribute when query

When doing query.order_by, sort_key_attr is get from model class,
we need to make sure sort_key_attr is really a QueryableAttribute
type instance before we do the query or it will cause errors.
This will prevent if there IS a function which name is same as sort_key.

Closes-Bug: 1405069
Change-Id: I8a3eb08ab3469ec08e05bfce754b664943d65c83