Updated openstack/openstack Project: openstack-dev/devstack  3eaa0781e3d87bee6d343573954b682f1f23d4b9
Fix get_packages when $DISTRO is not set

Sourcing the tools/install-prereqs.sh script with TOP_DIR set results
in GetDistro being called in get_packages and echoing the result.
Since all output from get_packages is assumed to be package names,
this results in the attempted installation of the non-existant 'Found'
'Distro' and '[distro name]' packages.  This change removes the echo
statement to avoid this problem.

Change-Id: Idd05c31b9eec9e6209666fa16fa425cdf1f35aa2