Updated openstack/openstack Project: openstack/neutron  2327f9f07e95cd0a5f35a90c4c7bfd783821b10a
Make L3 HA VIPs ordering consistent in keepalived.conf

Currently the order of VIPs in keepalived.conf is determined
by sorting the VIPs whenever one is added or removed. As it
turns out, keepalived doesn't like it when the primary VIP
changes. One side effect is that virtual routes, in our case
the router's default route, may be removed.

This patch fabricates an IP address on the router's HA interface
and uses it as the primary VIP.

Closes-Bug: #1404945
Change-Id: I993daf594a28918de6fafff465f5f40e7b89305e