Updated openstack/openstack Project: openstack/taskflow  b3e06a2e4688e8586212cbd8195f427754487ffe
Add split time capturing to the stop watch

For cases where it is useful to capture the elapsed time
of a watch and later examine those split times create a
method on the stop watch that allows for these kinds of
captures and iterations that correspond to the common stop
watch split capability.

Also adds basic docstrings to the stop watch public
methods so that people know what they are and can use them.

Change-Id: Ic52ae5dfcca9a117ccd0dda5cc62a14c09e15ce0