Updated openstack/openstack Project: openstack/neutron  051af204d9cbc61d012402a228b6b13f7c0878a2
Validate legacy router services before migration

When legacy routers are migrated to distributed
routers, we need to make sure that there are no
Advanced services associated with the router.

If Advanced services such as FWaaS or VPNaaS is
associated with the router, we raise an exception
and stop the migration.

Partial-bug: #1348309

DocImpact

Change-Id: I65909040a16b022108e481344064d375050d731c