Updated openstack/openstack Project: openstack/nova  4185b975ecf90886e65eae299aefd5f122c91167
Log sqlalchemy exception message in migration.py

The current code in migration.py suppress underlying exception message.
If Nova sql db configuration (eg. user, or permission) is incorrect, or
the query result is not as expected, nova-manage command will display
a confusing message, which may not help user to address the problem.
Need to log the exception message and stack trace, in order to improve
servicability.

Change-Id: I742f25dd643a7886a5fd5f66eee899ae5b17b439