Updated openstack/openstack Project: openstack/nova  c8bf2ea47a945116397123e0c9f34cf02a524cfc
XenAPI: Check image status before uploading data

Currently, the Xen API plugin sends the data payload to Glance
without checking the image status. In consequence, if the image is
not in queued status Glance will not let the image upload succeed.
This is by design, please refer [0].

The issue here being; buffering of the data will happen on the Glance
server. Consequently, the bound to be failed upload operation will not
get a 409 until after a large chunk of data has already been streamed
to the Glance server. This may result into client side wastage of
bandwidth and wait time.

This patch proposes a HEAD call be made from the glance plugin to
Glance API; thus in turn checking the valid image state for the
data transfer.

[0] http://docs.openstack.org/developer/glance/statuses.html

Change-Id: I24c1222a3b1ac809353724500117e58d50615ac8