Updated openstack/openstack Project: openstack/nova  4a7dbce02d1b6709e7ddeda88c17e4b6a1178d50
Add vision of nova rest API policy improvement in devref

There are several improvements for nova rest API policy. And we
should have a full view for them to make sure we are on the right way.

This docs aims to describe the improvement from blueprint v3-api-policy
and separated-policy-rule-v3-api.

Change-Id: I34a37475f2b18607fc9453764e675113c4ab773b