Updated openstack/openstack Project: openstack/neutron  a7b54d3c60ff6dd63273b573bcc5c572faab22a8
Fix topic for provider security group update

Commit 8098b6bd20bb12520e438ef9125dc7b96664f712 accidentally changed
the topic for the provider security group update to a regular member
update. This resulted in the L2 agent not asking for the latest
security group rules after a DHCP port was created. If a regular
compute port was brought online and wired up by the L2 agent
before the DHCP port was created, the VM would never get its allow
rule to communicate with the DHCP server.

Co-bug-squasher: Armando

Closes-Bug: #1403291
Change-Id: I382f2e1390c9a3ff558e152eeb76f0f827266cb4