Updated openstack/openstack Project: openstack/ceilometer  3c2f904efa52655fdd7ece208aaf208b0fa54e57
Remove unnecessary import_opt|group

Usually, we should import an option before we use it, otherwise,
NoSuchOpt exception will be raised.

When we import a module, the options in that module are registered,
(or imported), so we don't need to explicitly import them again.
However, it is a bit hard for developers, reviewers and maintainers
to track between files just to check if a specific option is imported
or not. So it is good (not required) whenever we use an option, we import
it explicitly, such kind of redundant is acceptable. But not the case that
we already import a module A, then import_opt|group registered or imported
in that module again, this is obviously unnecessary and should be avoid.

Change-Id: I8672d9f12d862f9a622e5557d9c8dc2fe6717f5e
Closes-Bug: #1389229