Updated openstack/openstack Project: openstack-infra/devstack-gate  537426a7a037beb1b52cd79b13e4832c3d7fa6b2
Copy tempest.log from verify-tempest-config

As part of the current devstack runs verify-tempest-config is run to
generate a complete extension list which is only used when a disable
extension option is provided. While this is not the best method for
doing this, since it exposes the risk of silently skipping tests if
there is an api bug it is currently how devstack supports doing this.
However, calling verify-tempest-config has one side-effect that it
generates a separate log file where it's run (at least using the
default oslo logging configuration) This commit adds a check if the
log file exists and copies it along with the other artifacts. So that
people will be able to debug failures in the verify-tempest-config
tool when they occur.

Change-Id: I76c78bcf1fe49903df7d0bcecbb4c81ea0d73ca2