Updated openstack/openstack Project: openstack/python-cinderclient  bae0bb3276f08b7b4a5441d8178a38e0dfc1e3e6
Add commands for modifying image metadata

These are Cinder Client specific changes for modifying volume
and snapshot's image metadata.

- cinder image-metadata volume_id set key=value
This command can be used to create a new image metadata associated
with the volume when the key not exists, and update the value for
the existed key.

- cinder image-metadata volume_id unset key
This command can be used to delete image metadata associated
with the volume.

Depends-On: I22792ef7bd49c763d7c130aa8cac9abc3fff2d4c
Partially implements: bp support-modify-volume-image-metadata
Change-Id: I03787563cb3ea4ef96ed145a37fc665f80565a12