Updated openstack/openstack Project: openstack/ceilometer  d81616bbae2ffbd4443fde0619b547e656e7dfbc
make transformers optional in pipeline

in the code we note that transformers do not need to be specified
but for some reason we have tests to ensure PipelineManager fails
if transformer is not specified. this patch adjust tests to make
sure PipelineManager does not fail if transformer is not specified.

Change-Id: I00c0e94da19d54c754bcf564d99eb439d2a820a1