Updated openstack/openstack Project: openstack/nova  25a546a1559568a41dd0b77a574b6953e4385fbc
Join instances column before expecting it to exist

test_create_instance_associates_security_groups was fetching a
security group without joining instances, then testing its instances
association. This causes the instances to be fetched after the db api
call has completed, and does not work when explicitly scoping a db
transaction to a db api call.

This change updates the test to explicitly fetch instances. We also
cleanup the test slightly, and add an additional assertion that the
retrieved group is the one we expected.

Change-Id: I11025d9ad13678760eae584951db22090a513645