Updated openstack/openstack Project: openstack-infra/subunit2sql  dd271423c25ec7f573c848d3d81841d32347accf
Convert tags metadata to string before persist it

Probably due to a bug on sqlalchemy's postgresql dialect set type wasn't
translated into a valid sql type when using postgresql.

This patch explicitly converts the tags set to a string separating the
elements with comma.

In mysql it ends up converted to a string containing the items and this
behavior is kept.

Change-Id: I27db345091fb4ae544cc4e778b001512361833d2
Story: 2000110