Updated openstack/openstack Project: openstack/nova  81533dc07e984dfb4c2997ca0bae5cfd59493cb2
libvirt: Handle empty context on _hard_reboot

When _hard_reboot runs for resume_guest_state_on_host_boot, we don't
have a valid context. This checks if the context is valid before
attempting to repopulate base images.

This is a bit of a hack, as it assumes that the base images already
exist on the system, and will fail if they don't. However, it reduces
the number of failures (without this, every downed guest is guaranteed
to error when coming back up for resume_guest_state_on_host_boot).

Closes-Bug: 1195884
Change-Id: Ibac52f5d1b4e5ab2adb0a561659e6df8c7c1f40f