Updated openstack/openstack Project: openstack/oslo.i18n  9a2cde208a46ea9cc46014ec0d276f382a2f9170
Add test fixture to prefix lazily translated messages

This fixture enables testing of lazy translation by faking out translations
by providing a translation that is the message id prefixed with the domain
and language used in the translation.  It supports the definition of new fake
languages so that a different language can be specified without having to
create the associated message catalogs.

For example, this will be used to test that a fake language requested using
the Accept-Language header is used to translate the message instead of
the default language.  That is, that the message was enabled for lazy
translation and was not translated until just before it was returned to the
user.

Change-Id: If6dd28815d11c07ce9430ed07e61bfbc23695707
Partially-Implements: blueprint i18n-enablement-juno