Updated openstack/openstack Project: openstack/cinder  c6d7e81ea937d8d42f50ffd20422865eaf979bcd
Garbage Remains when Attached Volume is Migrated with NFS Driver

Garbage volume remained when an attached NFS volume was migrated
to another NFS backend. An error message was generated as the
attached volume was not detached before being migrated. The root
cause of the issue is that status_update['status'] is only set
for retype and not for migration from one NFS backend to another
NFS backend. This patch proposes to fix the aforementioned issue
where migration fails from one NFS backend to another NFS backend.

Change-Id: I657a46d98b1e2a52cec128308cbbf39166c3625b
Closes-bug: 1391172