Updated openstack/openstack Project: openstack/cinder  0e9139ead2440d6e89ec208db9b3094dd7c5a6c4
Remove locks from LeftHand driver

By removing synchronization locks from the LeftHand driver,
support for many simultaneous volume create/deletes and
volume attach/detachs will be added.  This is useful for high
availability environments where performance with locks in place
would be reduced.

Removing the locks requires the creation of a new LeftHand
client when a request is sent to the driver now.  This
allows requests to no longer be blocked if a request is
already in process by the driver.

This results in a performance increase for the driver and
better stability in a high availability environment.

Closes-Bug: #1395953
Change-Id: I39732d944a79bbfc597077fe0873d4cfcaace211