Updated openstack/openstack Project: openstack/neutron  f8c92effe1e9509caebe4dc5d44b78eca5b275e4
Fixs shared networks in Arista ML2 driver

When a shared network is created, it is stored
in arista_provisioned_nets under the tenant_id of
owner of the network. Later, when a different
tenant launches an instance on the shared network,
above mentioned DB is looked to find the network
under requesting tenant's ID, and it is not found,
and hence, the request is rejected.

The fix is made such that when a network is not
found under the requesting tenant's ID, check if
this network is shared network. If yes, then
check the DB to ensure that owner tenant exists
in the DB. If yes, then this is a valid request
and do not reject it. Otherwise reject it - as an
unknown tenant or network is specified in the
request.

Change-Id: I8f25d8f84af844afdb3b607bd8ade32f0f5e81fb
Closes-bug: 1409176