Updated openstack/openstack Project: openstack/python-openstackclient  8dd8cdcdfebd0ad26bd3ddba8b6e75d1793d290b
Check if service.name available before access

Currently v3 endpoint commands access service.name directly, while
name is not a required attribute of service. So if we associate an
endpoint to a service without name, we will get an AttributeError
executing v3 endpoint commands later. This patch addresses this
issue by checking if service.name is available before accessing it.

Change-Id: I3dd686ef02a2e21e2049a49cb55634385c2ecfaf
Closes-Bug: #1406737