Updated openstack/openstack Project: openstack/nova  2b486a4e1af984ab3af8e2c018b04da0363c652d
Reduce the complexity of the create() method.

This is a refactor designed to reduce the cyclomatic complexity of the
current worst offender in nova: the servers.create() method of the nova
API. The complexity of the method was 46; it is now 22.

No logic changes were made.

This builds on the effort started with Change 129125 by Joe Gordon.

Closes-Bug: 1403586

Change-Id: Ib102091773f786b8c8dc5b79c6d776418ed1d57d