Updated openstack/openstack Project: openstack/heat  08e27758b99bb5ae09846890c5c81c4f1891f90a
Add support status to docs for properties

I thought we already had this, but inspection of the generated docs
reveals that we only do class-level docs for the support status,
not exposing those defined in the schema.

Adding the properties support status should help reduce user confusion
when new properties are added between releases.

Change-Id: I5f3228161e57daad73cbafc124c4fad05c68d3d6
Closes-Bug: 1376868