Updated openstack/openstack Project: openstack/os-net-config  6945fe5afdc12574fe00ad440319b7873f818e84
Add mapping option to influence nicN mapping order

Currently there's a fixed mapping between abstracted interface
names (nic1, nic2 etc) and the underlying biosdevname for the
device.

In many cases, this mapping based on system enumeration is
sufficient, but in some cases, particularly when you perform
detailed pre-deployment discovery of interfaces, you may wish
to alter the mapping independently of the config (e.g if the
config is in a heat template, and the discovery data is
provided at runtime).

So this adds a -m option to os-net-config, which enables a
mapping file to be provided, such that specific interfaces
may be mapped to their abstract names based on knowledge of
the devices or the networks they are connected to.

The mapping file has the following format, where em1 and em2 are
device names as detected by the OS (e.g biosdevname):

interface_mapping:
    nic1: em2
    nic2: em1

Or you can use the device MAC instead:

interface_mapping:
    nic1: 12:34:56:78:9a:bc
    nic2: 12:34:56:de:f0:12

Change-Id: I93e6d3ed733244834bb3c2126c91db705b4d9167