Updated openstack/openstack Project: openstack/keystone  ba3a58c0d24f09afeecef3fa880a8360933b5696
Split roles into their own backend within assignments

This is the first part of the more comprehensive split of
assignments, which rationalizes both the backend and controllers.
In order to make this change easier for reviewers, it is divided
into a number of smaller patches.

Follow-on patches will:

- Fix incorrect doc strings for grant driver methods
- Update unit tests to call the new role manager
- Update the assignment controller to call the role manager
- Refactor assignment manager and driver methods to logically
  separate project/domains from the actual assignments
- Split projects and domains into their own backend
- Split the controllers so they call the correct manager
- Update the tests to call the new correct manager

Partially implements: bp pluggable-assignments

Change-Id: I41fc23a049c26e514222a966c1847e183448be00